Add some type annotations
