+++ api/app/utils.py
@@ -3,13 +3,14 @@ import json
 from app.constants import BROADCAST_KEY
 from app.models import Event, Game
 from chess import Board
+from pika.channel import Channel
 
 
-def get_queue_name(gid, sid):
+def get_queue_name(gid: str, sid: str):
     return f"{gid}::{sid}"
 
 
-def get_redis_key(gid):
+def get_redis_key(gid: str):
     return f"game:{gid}"
 
 
@@ -17,7 +18,7 @@ def opponent_ind(turn: int):
     return int(not bool(turn))
 
 
-def serialise_game_state(game):
+def serialise_game_state(game: Game):
     """Serialise game state to JSON string for storage in Redis"""
     if not game:
         return
@@ -26,7 +27,7 @@ def serialise_game_state(game):
     return json.dumps(game_dict)
 
 
-def deserialise_game_state(game):
+def deserialise_game_state(game: str):
     """Deserialise game state from Redis JSON string"""
     if not game:
         return
@@ -35,6 +36,6 @@ def deserialise_game_state(game):
     return Game(**game_dict)
 
 
-def publish_event(channel, gid, event: Event, rk=BROADCAST_KEY):
+def publish_event(channel: Channel, gid: str, event: Event, rk=BROADCAST_KEY):
     # TODO: better place to put this?
     channel.basic_publish(exchange=gid, routing_key=rk, body=json.dumps(event.__dict__))
